+++ openwave/common/constants.py
@@ -37,15 +37,14 @@ RONTOSECOND = 1e-27  # s, rontosecond time scale
 # WAVE-MEDIUM
 # ================================================================
 MEDIUM_DENSITY = 3.859764604e22  # kg / m^3, wave-medium density (ρ)
+EWAVE_SPEED = 299792458  # m / s, energy-wave velocity (c, speed of light)
 
 # ================================================================
 # ENERGY-WAVE
 # ================================================================
-EWAVE_LENGTH = 2.854096501e-17  # m, energy-wave length (λ)
+EWAVE_FREQUENCY = 1.050393558e25  # Hz, energy-wave frequency (f = EWAVE_SPEED / EWAVE_LENGTH)
 EWAVE_AMPLITUDE = 9.215405708e-19  # m, energy-wave amplitude (A, equilibrium-to-peak)
-EWAVE_SPEED = 299792458  # m / s, energy-wave velocity (c, speed of light)
-
-EWAVE_FREQUENCY = 1.050393558e25  # Hz, energy-wave frequency (f = EWAVE_SPEED/EWAVE_LENGTH)
+EWAVE_LENGTH = 2.854096501e-17  # m, energy-wave length (λ = EWAVE_SPEED / EWAVE_FREQUENCY)
 EWAVE_PERIOD = 9.520241169e-26  # s, energy-wave period (T = 1 / EWAVE_FREQUENCY)
 
 # ================================================================